fix: fix enum
