--- src/types/enum.ts
@@ -1,16 +1,0 @@
-export enum ApiMethods {
-  GET = 'GET',
-  POST = 'POST',
-  DELETE = 'DELETE',
-  PUT = 'PUT',
-  PATCH = 'PATCH',
-}
-
-export enum ApiEndpoint {
-  AUTHENTICATION = '/oauth/token',
-}
-
-export enum ContentType {
-  URLENCODED = 'application/x-www-form-urlencoded',
-  JSON = 'application/json',
-}
