+++ src/types/enums.ts
@@ -15,3 +15,20 @@ export enum BUTTON_TYPE {
   SUBMIT = 'submit',
   RESET = 'reset',
 }
+
+export enum ApiMethods {
+  GET = 'GET',
+  POST = 'POST',
+  DELETE = 'DELETE',
+  PUT = 'PUT',
+  PATCH = 'PATCH',
+}
+
+export enum ApiEndpoint {
+  AUTHENTICATION = '/oauth/token',
+}
+
+export enum ContentType {
+  URLENCODED = 'application/x-www-form-urlencoded',
+  JSON = 'application/json',
+}
